[update] loader.py module into brkobj module
