+++ brkraw/api/brkobj/__init__.py
@@ -1,0 +1,4 @@
+from .study import StudyObj
+from .scan import ScanObj
+
+__all__ = [StudyObj, ScanObj]
